Refactor: Integration - Extract static css evaluation helpers
